tinycms analytics: geographic sessions component
